gpio: Configuring pinmux and power rail with gpio_request.
When  gpio client uses the direct linux api which is provided by
gpio lib of linux kernel driver, the pinmux and io power rail was
not getting configured.
Configuring the pinux and power rail so that client can use direct
gpio api of linux and need not to use the rm gpio wrapper api.

Tested on whistler.

Change-Id: I02b6782fd751543d064588288104a26caab7c848
Reviewed-on: http://git-master/r/1021
Reviewed-by: Suresh Mangipudi <smangipudi@nvidia.com>
Tested-by: Suresh Mangipudi <smangipudi@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>