tegra pmu: do not turn ON the LDO0 rail by default
The LDO0 power rail was enabled during bootup by default, thereby
increasing the ref count. This results in the rail remaining powered
ON even when we try to disable it from a driver.

Disabling the LDO0 by default, when there are no devices actively
using it, should result in power saving.

Change-Id: I19e478e9af2019de5e02ff9a31e7085e815ffaed
Reviewed-on: http://git-master/r/1058
Reviewed-by: Kenneth A Radtke <kradtke@nvidia.com>
Tested-by: Deepesh Subhash Gujarathi (Engrg-Mobile) <dgujarathi@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Tested-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>