tegra_odm_battery:porting the battery polling timer changes.
As we know that android doesn't polls for the battery status.
In order to update the battery and power supply related info
created a timer func which polls for every 30Sec and updates
the battery and power supply info.
Also, there is a sysfs attribure "status_poll_period" using
which this polling interval can be altered.

Already, we have these changes in nvec_battery.c which are
wokring and verified. Porting the same changes to
tegra_odm_battery.

Change-Id: I19f8e3f24c333745b7d90084a7e8c320ac8a1d6b
Reviewed-on: http://git-master/r/1064
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: Mayuresh Kulkarni <mkulkarni@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>
Tested-by: Sachin Nikam <snikam@nvidia.com>