fsl udc:Fix USB disconnect notification in OTG
In OTG mode, when USB cable is removed cable disconnect notification is not
sent to the client driver. Hence at application layer, UI shows up USB cable
connection notification even USB cable is removed.
With this fix disconnect notification is going to the client driver and
UI removes the USB notification on USB cable removal.

Tested on android/whistler

Bug 668692: USB cable is removed, but does not appear in notification.

Change-Id: I06d5a288873a2a8d7d109c5109b9dfd0c08fd1ee
Reviewed-on: http://git-master/r/1112
Reviewed-by: Gary King <gking@nvidia.com>
Tested-by: Gary King <gking@nvidia.com>