panjit touch: Adding the odm panjit touch screen driver.
Checking in odm panjit touch screen driver. This is required for
new harmony tablet board. This is disabled by default as of now.
Also checking updated tegra-gpio.c to have the correct gpio lies
info for harmony tablet. This is also disabled by default.
Added the acceleremeter properties in the odm kit. The driver
for accelerometer will be checked in later.
updated copyright headers for panjit.

Change-Id: I7089ddb11670670dc0aa683dce190bc619fb07e2
Reviewed-on: http://git-master/r/1139
Reviewed-by: Gary King <gking@nvidia.com>
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Tested-by: Krishna Reddy <vdumpa@nvidia.com>