tegra usb: fine tuning CPU clock boost freq for USB device
The CPU clock was previously boosted to 300 MHz for USB device on cable
connect for performance optimization. Following issues were noted with
this setting:
1. it is above low limit (250MHz) used to turn off slave CPU. Hence it
will never be turned off when cable is connected.
2. it is also above PLLP 216MHz output. Hence high power PLLX will be
kept alive.

Reducing the CPU boost frequency to 200 MHz to address the above issues
with minimal impact on the performance.

Bug 654486: [whistler/android] - Large difference in USB MSD perf with
DFS OFF.

Change-Id: Ia57db137b493dd8fe8436ede7942df0661dad67b
Reviewed-on: http://git-master/r/1228
Tested-by: Abhishek Aggarwal <aaggarwal@nvidia.com>
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>