tegra ALSA : Handle Underrun Case
In case of underrun application triggers stop and starts
the playback again, don't exit the thread in case of stop.
underrun can happen when an application does not feed new
samples in time to alsa-lib (due to CPU usage). With this
change breaks may be observed in case of underrun which
cannot be avoided, but the application will finish the
playback and exits gracefully.

For bug:663146 [LDK/Chromium/ALSA/Stability] "underrun!!!"
(pcm_write:1442: write error: Input/output error) issues
in ALSA while playback.

Change-Id: Ib9b30461b4e381a66c3fd48467406e448aa41a17
Reviewed-on: http://git-master/r/1257
Reviewed-by: Vijay Mali <vmali@nvidia.com>
Tested-by: Vijay Mali <vmali@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>