tegra ALSA: Do not queue buffers to the mixer in stopped state
In SNDRV_PCM_TRIGGER_STOP state of the driver, if a buffer gets
queued then it might not be returned back as it never gets rendered.
In this scenario play_thread never exits as it keep on waiting for
receiving all the buffers back from mixer before exiting.

This is to solve bug 684393.

In case of START->STOP->START scenario (happens in case of underruns),
without closing the stream, there will not be any buffer loss as buffers
are queued by application at appl_ptr and they remain there, once next
START is called, they will be picked up for the playback.

Change-Id: I4536e35bc59a6a4d67b24d9a8dc57a738ec695e3
Reviewed-on: http://git-master/r/1323
Reviewed-by: Gary King <gking@nvidia.com>
Reviewed-by: Vijay Mali <vmali@nvidia.com>
Tested-by: Vijay Mali <vmali@nvidia.com>
Reviewed-by: Sandeep Shinde <sashinde@nvidia.com>
Tested-by: Sandeep Shinde <sashinde@nvidia.com>