tegra ehci: Shutdown ehci only when host is in operation
During ehci shutdown, the driver accesses the controller registers and
hence it powers up the phy. It is observed that during re-powering as
part of shutdown, there is a hang while accessing ULPI view_port register.

Preventing this by modifying ehci shutdown code to shutdown only when
the ehci is in operation.

Bug 685566

Change-Id: If69a23071c56c8f76c6f3c6615c1b948a452b9ce
Reviewed-on: http://git-master/r/1397
Reviewed-by: Hanumanth Venkateswa Moganty <vmoganty@nvidia.com>
Tested-by: Nitin Ghate <nghate@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>