tegra mmc: enable MMC_UNSAFE_RESUME config for tegra_harmony_android
Normally, when the system goes into suspend, all the storage cards are
removed and later re-enumerated upon resume. During re-enumeration the
device name and number are not guaranteed to be the same as they were
during bootup. This can cause data corruption on cards.

When the MMC_UNSAFE_RESUME config is enabled, it will prevent kernel
from unmounting the mmc cards before suspend. Upon resume the cards
will still be using the same device id as they were before suspend.

This change seems to fix a lot of mmc related issues that occur
during LP0 transitions.

partial fix for Bug 679959

Change-Id: Ib90f95ebbc81c5b6e9e83e740ece31f914a262eb
Reviewed-on: http://git-master/r/1412
Reviewed-by: Narendra Damahe <ndamahe@nvidia.com>
Tested-by: Narendra Damahe <ndamahe@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>