tegra framebuffer - dsi one shot support
DSI one shot support requires register access in the tegra framebuffer to
poke the frame trigger bit (with both the trigger bit and the tearing effect
signal are high, a frame of pixels will be sent to the panel). The boot args
must also be expanded to have a "use tearing effect" flag.

Change-Id: Id0c3960cac5b59dd23fd6547f1b7dae4d422ac32
Reviewed-on: http://git-master/r/1460
Reviewed-by: Gary King <gking@nvidia.com>
Tested-by: Arthur Spence <aspence@nvidia.com>