tegra serial: adding power control api for bluesleep power management.
Adding apis for following function to hooking up tegra serial driver with
bluesleep power management:
 - Clock off.
 - Clock on.
 - Setting flow control to desired state.
 - Checking for tx fifo status.

Following 4 state machine is developed to achieve this:
UART_CLOSED, UART_OPENED, UART_SUSPEND, UART_CLOCK_OFF.

The transitions of state machines are as follows:
UART_CLOSED: the init state  on which resource is allocated but not
  opened by client or device is closed.

UART_OPENED: Able to do data transfer.
  CLOSED  to  OPENED by opening the port.
  CLOCK_OFF to OPENED by calling function tegra_uart_request_clock_on().
  SUSPEND to OPENED by calling resume().

UART_CLOCK_OFF: The controller clock is disabled and so no data transfer
 can happen. At this state, controller is not ready for deep power down.
 OPENED to CLOCK_OFF by calling tegra_uart_request_clock_off().
 Can not go to this state from CLOSED and SUSPEND.

UART_SUSPEND: The controller is in suspended state and ready for deep power down.
 UART_CLOCK_OFF to SUSPEND:
 OPENED to SUSPEND.

Tested in whistler.

Change-Id: I24461ac6f20f61e44b49332cdc6b93a01f605fc9
Reviewed-on: http://git-master/r/1377
Reviewed-by: Udaykumar Rameshchan Raval <uraval@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>
Tested-by: Gary King <gking@nvidia.com>