tegra sdhci: Enabling sdio cards suspend/resume
SDIO-WIFI Behavior -
1) With or without MMC_UNSAFE_RESUME setting -
As "sdio_driver" in mmc stack doesn't have suspend/resume handlers
so even if MMC_UNSAFE_RESUME is set, SDIO-WIFI driver's remove/probe
is called for mmc suspend/resume.
Thus, on suspend (LP1 or LP0) wifi is turned off and is disconnected.
On resume (from LP1 or LP0), wifi is turned on and is reconnected.

2) For all cards including sdio, power to rail is cut.
On suspend for sdio slot NvOdmSdioSuspend() is called.
On resume for sdio slot NvOdmSdioResume() is called.

Change-Id: I52e65b08a884a16f47cf5b1e4fcbb187ea4281e6
Reviewed-on: http://git-master/r/1776
Reviewed-by: Kapil Hali <khali@nvidia.com>
Reviewed-by: Pavan Kunapuli <pkunapuli@nvidia.com>
Reviewed-by: Rahul Bansal <rbansal@nvidia.com>
Tested-by: Rahul Bansal <rbansal@nvidia.com>
Reviewed-by: Narendra Damahe <ndamahe@nvidia.com>
Tested-by: Narendra Damahe <ndamahe@nvidia.com>
Tested-by: Rakesh Goyal <rgoyal@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>