video: pwm_backlight: Add check_fb hook
In systems with multiple framebuffer devices, one of the devices might
be blanked while another is unblanked.  In order for the backlight
blanking logic to know whether to turn off the backlight for a
particular framebuffer's blanking notification, it needs to be able to
check if a given framebuffer device corresponds to the backlight.

This plumbs the check_fb hook from core backlight through the
pwm_backlight helper to allow platform code to plug in a check_fb hook.

Originally reviewed on http://git-master/r/21716

Change-Id: I2b2530e515227eaba246750b3cee7e61e882d14e
Reviewed-on: http://git-master/r/23100
Tested-by: Gaurav Sarode <gsarode@nvidia.com>
Reviewed-by: Robert Morell <rmorell@nvidia.com>
Reviewed-by: Mayuresh Kulkarni <mkulkarni@nvidia.com>
Reviewed-by: Daniel Willemsen <dwillemsen@nvidia.com>