ARM: tegra: ventana: Fixed backlight toggling.
Ventana has both an internal panel (on fb0) and external HDMI (on
fb1). One of the devices might be blanked while another is unblanked.
In order for the backlight blanking logic to know whether to turn off
the backlight for a particular framebuffer's blanking notification, it
needs to be able to check if a given framebuffer device corresponds to
the backlight.

This implement the check_fb pwm_backlight hook to make the internal
panel's backlight only toggle on fb0 blank changes.

Making equivalent commit from http://git-master/r/21717

Fixed bug 798863

Change-Id: I89827f06d71500bb48366b9c68fea1fe14508bb3
Reviewed-on: http://git-master/r/23101
Reviewed-by: Varun Colbert <vcolbert@nvidia.com>
Tested-by: Varun Colbert <vcolbert@nvidia.com>