arm: tegra: clock: Reading APB bus before disabling clock
It may be possible that write operation on apb bus does not get
complete before disabling clock if the clock is disabled just after
the write on apb bus.
To have proper sequence of operation, it is require to read back the
apb bus to make sure the write operation is completed.

bug 830481

Change-Id: I7f9f68f4cd6d39cf0bd697ddd236c4ce733dcf43
Reviewed-on: http://git-master/r/34413
Reviewed-by: Niket Sirsi <nsirsi@nvidia.com>
Tested-by: Niket Sirsi <nsirsi@nvidia.com>