nvhost: Do not crash if wait syncpt > max
If user space is waiting for a syncpt value higher than max,
nvhost_wait_timeout() panics the kernel. This panic is based on the
assumption that nobody should wait for sync points higher than indicated
by a submit from user space.

As the API has nothing to disallow waiting for future sync points not
indicated by any submit, this patch removes the panic and treats this as
a normal case.
Bug 843238

Change-Id: I367c46e42bd683f1023e7fe04e523a99ab3d666b
Reviewed-on: http://git-master/r/36470
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-by: Scott Williams <scwilliams@nvidia.com>
Tested-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>