sdhci: tegra: Read clk ctrl reg before clk_disable
It may be possible that write operation on ahb bus does not get
complete before disabling clock if the clock is disabled just after
the write on ahb bus.
To have proper sequence of operation, it is require to read back some
register to make sure the write operation is completed.

Bug 833340

Change-Id: Ib75ba95d86e028e9a08aa8561662d2e643873e60
Reviewed-on: http://git-master/r/39166
Tested-by: Pavan Kunapuli <pkunapuli@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>