arm: tegra: clock: Reading APB bus before disabling clock
It may be possible that write operation on apb bus does not get
complete before disabling clock if the clock is disabled just after
the write on apb bus.
To have proper sequence of operation, it is require to read back the
apb bus to make sure the write operation is completed.

bug 833341

This is cherry pick of change http://git-master/r/#change,32556
in main but hand-merged.

Change-Id: I1d73a2fdfb35450220f53973d56984c9b89fdd27
Reviewed-on: http://git-master/r/45985
Reviewed-by: Alok Chauhan <alokc@nvidia.com>
Tested-by: Alok Chauhan <alokc@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: Mayuresh Kulkarni <mkulkarni@nvidia.com>