tegra: usb: ehci: synchronize shutdown and resume
The system is likely to hang if user issues "reboot" command
when a user space program is accessing USB device. In this
case, the reboot command will cause tegra_ehci_shutdown to be
invoked, at the same time, the tegra_ehci_bus_resume may be
running and accessing registers when it's not powered on.

bug 798849

Change-Id: I771a1f5fc61d95c36e058b61dd88d146ff948dee
Reviewed-on: http://git-master/r/47340
Reviewed-by: Bo Yan <byan@nvidia.com>
Tested-by: Bo Yan <byan@nvidia.com>
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>