mfd: fix tps6586x GPIO value setting
The tps6586x_gpio_set function is resetting all other GPIO pins.
We need to update the right GPIOxOUT bit of the GPIOSET2 register instead
of overriding the full value.

Also Inverted GPIO and subdevices initialization.
When using a fixed voltage regulator, this allows to declare
and initialize it conveniently from the "subdev" list.

Bug: 829647
Change-Id: I1480c2d1f2bee549ebc81fef264841c9a9e4daaa
signed-off-by: Preetham Chandru <pchandru@nvidia.com>
Reviewed-on: http://git-master/r/47809
Reviewed-by: Allen R Martin <amartin@nvidia.com>