Ventana: KBC: Removing the KBC usage on ventana
On Ventana, Keys are handled through the GPIO driver. So CONFIG_KEYBOARD_TEGRA
should not be defined

On LDK, kernel defconfig of Harmony is shared with Ventana(to move towards
SingleBinary for multiple tegra boards). Since on Harmony we have
to support the matrix keyboard through KBC, we have enabled the KBC driver by
defining CONFIG_KEYBOARD_TEGRA which reflects in Ventana also.
This situation leads to an spurious interrupt from KBC as soon as system
enters into the LP0 state.

To avoid the early exit from LP0, we are removing the KBC device registration
on board_ventana.c

Bug 869778

Change-Id: I6a097ad8713d2648fd08f0657a37b826e47c2379
Reviewed-on: http://git-master/r/52049
Reviewed-by: Ramalingam C <ramalingamc@nvidia.com>
Tested-by: Ramalingam C <ramalingamc@nvidia.com>
Reviewed-by: Victor Ryabukhin <vryabukhin@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>