video: tegra: avp: set emc to max if avp clock is max
this change ensures that emc clock is set to maximum when
avp clock goes maximum. it over-rides the emc clock
send by platform data when avp clock is maximum. it restores
the platform specific emc clock (if any) when the next
non-maximum avp clock request comes. if no emc platform
specific clock is specified, it restores the maximum emc clock

for bug 876402

Change-Id: Id5e454a5a63b1679906d8612c6e24958bace1663
Signed-off-by: Mayuresh Kulkarni <mkulkarni@nvidia.com>
Reviewed-on: http://git-master/r/53193
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Tested-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: Mohan Nimaje <mnimaje@nvidia.com>