ov5650: add fast set_mode
Resetting the mode does not need to reset all of the i2c
registers. This used to be called "fast set_mode". This
is added into the kernel driver in this change. This
change reduced 90ms from resetting the mode. It only
apply to still preview and capture resolutions

bug 816814

Change-Id: Id27f526dc8b4623b08a04b536d8d7ebdc904572d
Reviewed-on: http://git-master/r/53587
Tested-by: Anton Poon <antonp@nvidia.com>
Reviewed-by: Qi Wang <qiw@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>