video: tegra: dc: ignore unused overlays for bandwidth
If an overlay is not being used, do not program the latency allowance.

This is to avoid underflows that occur at a resolution of 19x12.  When the
unused overlays are reenabled, they underflow if the latency allowance has
previously been increased to a very high value.

(cherry picked from commit 8a4c47b17fae10a65e4816e419dff46b9f4785d1)

Change-Id: I825b4c1659a9f4f982bc66513b08b95879f17dd5
Reviewed-on: http://git-master/r/62522
Reviewed-by: Lokesh Pathak <lpathak@nvidia.com>
Tested-by: Lokesh Pathak <lpathak@nvidia.com>